+++ eval.py
@@ -24,27 +24,26 @@ def load_labels(fn):
 
 
 def evaluation(matcher, trajs_folder, debug_folder=None):
-
     trajs = trajs_folder.glob("*.geojson")
     gt_fn = trajs_folder / 'gt.json'
     labels = load_labels(gt_fn)
 
     if debug_folder is None:
         debug_folder = DATA_FOLDER / "result"
-    debug_folder.mkdir(exist_ok = True)
-    
+    debug_folder.mkdir(exist_ok=True)
+
     preds = {}
     hit = 0
     errors = {}
     timer = Timer()
     timer.start()
-        
+
     for fn in tqdm(trajs):
         name = fn.name
         traj = matcher.load_points(fn, compress=False)
-        path, i = matcher.matching(traj, plot=False, top_k=5, dir_trans=True, 
-                                   debug_in_levels=False, save_fn = debug_folder / str(name).replace('geojson', 'jpg'))
-        y = path.eid.values
+        res = matcher.matching(traj, plot=False, top_k=5, dir_trans=True,
+                               debug_in_levels=False)  # , save_fn = debug_folder / str(name).replace('geojson', 'jpg'))
+        y = res['eids']
         # preds[fn.name] = [int(i) for i in y]
 
         if np.array(y == labels[name]).all():
@@ -52,11 +51,12 @@ def evaluation(matcher, trajs_folder, debug_folder=None):
         else:
             errors[name] = fn
 
-    print(f"Prcision: {hit / (hit + len(errors)) * 100:.1f} %, time cost: {timer.stop():.2f} s")
+    print(
+        f"Prcision: {hit / (hit + len(errors)) * 100:.1f} %, time cost: {timer.stop():.2f} s")
     if len(errors):
         print(f"Errors: {errors.keys()}")
-        
-    return 
+
+    return
 
 
 if __name__ == "__main__":