+++ mapmatching/__init__.py
@@ -1,2 +1,2 @@
 from .utils.timer import Timer, timeit
-from .matching import build_geograph, ST_Matching
+from .matching import build_geograph, ST_Matching, STATUS