+++ mapmatching/match/projection.py
@@ -9,7 +9,6 @@ from ..graph import GeoDigraph
 def project_traj_points_to_network(traj_panos:gpd.GeoDataFrame, path:gpd.GeoDataFrame, net:GeoDigraph, keep_attrs=[['pid', 'geometry']]):
     panos = traj_panos.copy()
     panos.loc[:, 'eid'] = panos.apply(lambda x: path.loc[path.distance(x.geometry).idxmin()].eid, axis=1)
-    # panos.loc[:, 'rid_root'] = rid_root
 
     projectd_edges = net.df_edges.loc[panos['eid'].values]
     edge_geoms = projectd_edges.geometry.values
@@ -23,7 +22,6 @@ def project_traj_points_to_network(traj_panos:gpd.GeoDataFrame, path:gpd.GeoData
         keep_attrs = list(panos)
     panos = pd.concat([panos[keep_attrs], df_project], axis=1)
     panos.sort_values(['eid', 'offset'], inplace=True)
-    # plot_geodata(gpd.GeoDataFrame(panos).set_geometry('projected_point'))
 
     return panos